Easier runtime config for mapping mocks and remove unnecessary mock cloud adapter config (#52)
Makes it easier to configure the mapping mocks (both in-memory and external) at runtime with layered configuration files. Also removes the mock cloud adapter config which was not used.

This addresses two of the tasks in #17, but more work is needed before that can be closed. This PR should serve as a precedent for the remaining work.